+++ include/utils/pixel.h
@@ -12,7 +12,7 @@ void bgr2rgb24(unsigned char *image, unsigned int w, unsigned int h);
 void bgra2rgba32(unsigned char *image, unsigned int w, unsigned int h);
 void argb2rgba32(unsigned char *image, unsigned int w, unsigned int h);
 
-unsigned char *scaleBuffer(unsigned char *image, int width, int height, int components);
+unsigned char *scaleBuffer(unsigned char *image, unsigned int *w, unsigned int *h, unsigned int bpp);
 
 #endif